Rename `mounted-datasets` to `datasets-in-use`
