+++ pkg/webhook/plugins/plugins_impl.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/fluid-cloudnative/fluid/pkg/common"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/api"
+	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/datasetusageinjector"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/fusesidecar"
-	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/mounteddatasetinjector"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/mountpropagationinjector"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/nodeaffinitywithcache"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/prefernodeswithoutcache"
@@ -49,7 +49,7 @@ func RegisterMutatingHandlers(client client.Client) error {
 	_ = registry.Register(requirenodewithfuse.Name, requirenodewithfuse.NewPlugin)
 	_ = registry.Register(nodeaffinitywithcache.Name, nodeaffinitywithcache.NewPlugin)
 	_ = registry.Register(fusesidecar.Name, fusesidecar.NewPlugin)
-	_ = registry.Register(mounteddatasetinjector.Name, mounteddatasetinjector.NewPlugin)
+	_ = registry.Register(datasetusageinjector.Name, datasetusageinjector.NewPlugin)
 
 	// get the handlers through the config file
 	data, err := os.ReadFile(common.WebhookPluginFilePath)
